# ddtrace/tracer: skip creating Meta and Metrics map when not needed