This provides a small performance improvement on creating spans.

     benchmark                old ns/op     new ns/op     delta
     BenchmarkStartSpan-8     574           501           -12.72%

The call to `runtime.makemap_small` is eliminated in cases where the `Meta` and `Metrics` maps are not needed at start time.