@erichgess 2019-10-25
So the gain was from removing the initialization of the metric and meta maps in the constructor and only initialize if a metric/meta field is added?  Very smart.